Refactor layout calculations using dataclasses
