+++ src/model/__init__.py
@@ -2,5 +2,6 @@
 # Contains data structures and business logic independent of UI
 
 from model.struct_model import StructModel, parse_struct_definition, calculate_layout
+from model.layout import LayoutCalculator, LayoutItem
 
-__all__ = ['StructModel', 'parse_struct_definition', 'calculate_layout'] 
+__all__ = ['StructModel', 'parse_struct_definition', 'calculate_layout', 'LayoutCalculator', 'LayoutItem']
